test0/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.7
spacy==3.7.2
python-dotenv==1.0.0
weasyprint==60.2
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
